requirements.txt
faster-whisper
TTS
torch
torchvision
torchaudio
livekit-api
livekit-agents
PyJWT
numpy
scipy
sounddevice
protobuf==3.20.3
python-dotenv
requests
aiohttp
webrtcvad
pyaudio
notebook